+++ src/main/java/com/shop/demo/coupons/AccountCoupon.java
@@ -1,0 +1,29 @@
+package com.shop.demo.coupons;
+
+import lombok.*;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+@Entity
+@Getter
+@AllArgsConstructor
+@EqualsAndHashCode(of = "id")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class AccountCoupon {
+
+    // TODO: 2020/10/07 대대적인 공사필요
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    @Column(nullable = false)
+    private Long accountId;
+
+    @Column(nullable = false)
+    private Long couponId;
+
+    private boolean isUsed;
+}
