--- src/main/java/com/shop/demo/domain/questions/Answer.java
@@ -1,22 +1,0 @@
-package com.shop.demo.domain.questions;
-
-import lombok.*;
-
-import javax.persistence.*;
-
-@Entity
-@Getter
-@AllArgsConstructor
-@EqualsAndHashCode(of = "id")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Answer {
-
-    @Id  @GeneratedValue
-    private Long id;
-
-    private String answer;
-
-    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.REMOVE)
-    private Question question;
-
-}
