+++ src/main/java/com/shop/demo/orders/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.shop.demo.domain.orders;
+package com.shop.demo.orders;
 
 public enum OrderStatus {
     입금완료, 주문완료, 주문취소
