+++ src/main/java/com/shop/demo/orders/ShippingStatus.java
@@ -1,4 +1,4 @@
-package com.shop.demo.domain.orders;
+package com.shop.demo.orders;
 
 import lombok.Getter;
