+++ src/main/java/com/shop/demo/products/Product.java
@@ -1,4 +1,4 @@
-package com.shop.demo.domain.products;
+package com.shop.demo.products;
 
 import com.shop.demo.common.Money;
 import lombok.*;
